--- locker/templates/my_loans.html
@@ -1,15 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}My Loans{% endblock %}
-
-{% block content %}
-
-<div class="table-container">
-<h1>Locker</h1>
-<ul>
-    <li><a href="{% url 'item_list' %}">Items</a></li>
-    <li><a href="{% url 'checkout' %}">Checkout</a></li>
-    <li><a href="{% url 'return_items' %}">Return</a></li>
-</ul>
-
-{% endblock %}
