+++ static/css/pba.css
@@ -1833,53 +1833,3 @@ body {
     gap: 1rem;
   }
 }
-
-
-/* button styling for item returns*/
-.button-checkbox-list {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.button-checkbox {
-  display: inline-block;
-  margin: 5px;
-  width: 100%;
-}
-
-.button-checkbox input[type="checkbox"] {
-  display: none !important;
-  /* Remove any form-input styling */
-  border: none;
-  padding: 0;
-  margin: 0;
-}
-
-form .button-checkbox label,
-.button-checkbox label {
-  display: inline-block;
-  padding: 10px 20px;
-  background-color: #f0f0f0;
-  border: 2px solid #ccc;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: all 0.3s;
-
-  text-align: center;
-  width: 100%;
-  height: auto;
-  font-size: inherit;
-}
-
-form .button-checkbox input[type="checkbox"]:checked + label,
-.button-checkbox input[type="checkbox"]:checked + label {
-  background-color: rgb(131, 189, 86);
-  color: white;
-  border-color: rgb(131, 189, 86);
-}
-
-form .button-checkbox label:hover,
-.button-checkbox label:hover {
-  background-color: #e0e0e0;
-}
